# target/classes/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Account Management API Documentation
  description: API documentation for the Account Management System
servers:
  - url: http://127.0.0.1:8080
    description: Local server
paths:
  /customers:
    post:
      summary: Create a new customer
      operationId: createCustomer
      description: Create a new customer
      tags:
        - Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        201:
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        400:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
    get:
      summary: Retrieve all customers
      operationId: getAllCustomers
      description: Retrieve all customers
      tags:
        - Customer
      responses:
        200:
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /customers/{id}:
    get:
      summary: Retrieve a customer by ID
      operationId: getCustomerById
      description: Retrieve a customer by ID
      tags:
        - Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Update customer information
      operationId: updateCustomer
      description: Update customer information by ID
      tags:
        - Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        200:
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        400:
          $ref: '#/components/responses/ErrorResponse'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Delete a customer by ID
      operationId: deleteCustomer
      description: Delete a customer by ID
      tags:
        - Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        204:
          description: Customer deleted successfully
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
...
  /accounts/customer/{customerId}:
    get:
      summary: Retrieve accounts by customer ID
      operationId: getAccountsByCustomerId
      description: Retrieve accounts by customer ID
      tags:
        - Account
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Accounts for the customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
...
  /customerStatuses:
    get:
      summary: Retrieve all customer statuses
      operationId: getAllCustomerStatuses
      description: Retrieve all customer statuses
      tags:
        - Status
      responses:
        200:
          description: A list of customer statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerStatus'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /customerStatuses/{id}:
    get:
      summary: Retrieve a customer status by ID
      operationId: getCustomerStatusById
      description: Retrieve a customer status by ID
      tags:
        - Status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Customer status details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerStatus'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        dateOfBirth:
          type: string
          format: date
        statusId:
          type: integer
          format: int64
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1234
        accountNumber:
          type: string
        accountTypeId:
          type: integer
          format: int64
        balance:
          type: number
          format: double
        customerId:
          type: integer
          format: int64
        statusId:
          type: integer
          format: int64
...
    CustomerStatus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        statusName:
          type: string
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string